handle PR synchronize (ie new commits pushed)
When someone pushes commits to a branch that has a PR up, the PR will
"synchronize" itself against the branch, which basically means the PR
picks up those new pushes. When that happens, post to the PR thread.

Also, some refactoring: I pulled out "check if we have a PR thread, and
if so act on it" logic out from `handlePrClosed`, since the new
`handlePrSynchronized` has the same boilerplate (and others will as
well).