always notify on approvals and request-changes
Approvals and requests for changes are always important. Only dedupe the
slack notifications for general comments. But, always _set_ the
lastCommented. So:

* if someone approves and then comments, you'll get a single message:
  "$Person approved (possibly with comments)"
* if someone comments and then approves, you'll get two messages:
  "$Person commented" and then "$Person approved (possibly with
  comments)"

Also, remove the "possibly with comments" from the "requested changes"
notification. If they requested changes, we can probably assume they
left at least one comment saying why! The "(possibly with comments)" is
really just a reminder to the PR author to look at the comment thread
before blindly merging the PR, which isn't a concern when someone
requests changes.