configure the app when we enter/leave a room
Add slack event listeners that keep track of which channels the bot has
been added to. Rather than hard-coding the channel id, we now just store
the channels we've been added to in a `@klotho::persist`.

Supporting changes:

- Split the routers up from index.ts so that we can put the
  slack-listener (new) and the github-listener (which was there
  previously) in different exec units.
- Create a slack-common.ts for various things the two listeners need,
  including the secrets; this makes for a nice topo map.

This fixes #3 (though it doesn't include the `/list` functionality that
ticket mentions).