handle pings
With the introduction of the slack integration, we now fail to respond
to the github ping until we've set up slack: we try to get the botId,
and that causes us to get the slack creds, and that causes us to fail.

We don't need the botId at that stage, so let's handle the ping
separately. To make this work well, we also lazy-load the `Slack` obj.

In the future, that will also let us hook up a secret if we want.