make emoji customizable
This involves two main steps: first, making an EmojiStore to record
which emoji we want for each action; and second, adding slack commands
and actions for setting these.

As part of this, I'm also changing the default emoji to standards ones.
This makes it work out of the box for people installing it without our
custom emoji.

Resolves #37